refactor: made payment integration modular
